SampleEditor: use long long for target frames
for both the length in frames of the resulting sample and the real time
transport position we do now use `long long` instead of `int` and
`unsigned long`. Else we could get in trouble with very long samples or
long lasting Hydrogen sessions in which a lot of sample editing is done
but transport won't be started in a while.

+++ src/gui/src/SampleEditor/DetailWaveDisplay.cpp
@@ -115,7 +115,7 @@ void DetailWaveDisplay::drawPeakData()
 	p.setRenderHint( QPainter::Antialiasing );
 	const int nVerticalCenter = height() / 2;
 
-	int nnFrame;
+	long long nnFrame;
 	switch ( m_pSampleEditor->getSelectedSlider() ) {
 		case SampleEditor::Slider::Start:
 			nnFrame = m_pSampleEditor->getLoopStartFrame();
